# docs/openapi/agent-access-v1.yaml
openapi: 3.0.3
info:
  title: ActWeave Agent Access Protocol
  version: '2026-07-20'
  description: >-
    Workspace- and Agent-scoped API for service principals. Management user JWTs
    are not valid on this surface. Persisted Run events use the AAP v1 Event
    Envelope; SSE transport signals are not persisted events.
servers:
  - url: /api/agent-access/v1
security:
  - AAPAccessToken: []
paths:
...
  /workspaces/{workspaceId}/agents/{agentId}/profile:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/AgentId'}
    get:
      operationId: getAgentProfile
      responses:
        '200':
          description: Public integration profile for the bound Agent.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AgentProfile'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/NotVisible'}
...
components:
  securitySchemes:
    AAPAccessToken:
      type: http
      scheme: bearer
      bearerFormat: AAP JWT
      description: Short-lived token with Agent Access audience; user JWTs are rejected.
  parameters:
    WorkspaceId: {name: workspaceId, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
    AgentId: {name: agentId, in: path, required: true, schema: {$ref: '#/components/schemas/UUID'}}
...
  schemas:
    UUID: {type: string, format: uuid}
...
    AgentProfile:
      type: object
      additionalProperties: false
      required: [object, id, name, description, version, supportedContent, capabilities, interactionRequirements]
      properties:
        object: {type: string, enum: [agent_profile]}
        id: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        description: {type: string}
        version: {type: string}
        supportedContent: {type: array, items: {type: object}}
        capabilities: {type: array, items: {type: object}}
        interactionRequirements: {type: object}
...
    ErrorValue:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorValue'}
  responses:
...
    Unauthenticated:
      description: Missing, expired, or invalid Agent Access token.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotVisible:
      description: Resource does not exist or is not visible in the bound Scope.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...